# api/app/signals/apps/api/templates/api/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Signalen Informatievoorziening Amsterdam (SIA) API spec V1"
  description: >-
    One of the tasks of the municipality of Amsterdam is to maintain
    public spaces like roads, parks. When citizens have complaints about public
    spaces they can leave these complaints with the municipality. SIA, the
    Signalen Informatievoorziening Amsterdam, receives these complaints and is
    used to track progress towards their resolution.


    SIA provides an API that is used both by the SIA frontend and external
    systems that integrate with SIA. This swagger file documents that API.
  contact:
    email: "datapunt@amsterdam.nl"
  version: "TBD"
paths:
...
  /signals/v1/private/signals/geography:
    get:
      description: Signals geography list endpoint
      parameters:
        - name: "created_before"
          in: query
          description: >-
            Filter returns only signals created before specified datetime. Note
            the date time string should be in ISO 8601 format and URL encoded.
          schema:
            type: string
            format: date-time
          required: false
        - name: "created_after"
          in: query
          description: >-
            Filter returns only signals created after specified datetime. Note
            the date time string should be in ISO 8601 format and URL encoded.
          schema:
            type: string
            format: date-time
          required: false
        - name: "updated_before"
          in: query
          description: >-
            Filter returns only signals updated before specified datetime. Note
            the date time string should be in ISO 8601 format and URL encoded.
          schema:
            type: string
            format: date-time
          required: false
        - name: updated_after
          in: query
          description: >-
            Filter returns only signals updated after specified datetime. Note
            the date time string should be in ISO 8601 format and URL encoded.
          schema:
            type: string
            format: date-time
          required: false
        - name: status
          in: query
          description: Filter returns only signals with specified status.
          schema:
            $ref: '#/components/schemas/StatusStateChoices'
          required: false
        - name: maincategory_slug
          description: Filter returns only signals with matching main category
            slug.
          in: query
          required: false
          schema:
            type: string
        - name: category_slug
          description: >-
            Filter returns only signals with matching sub category slug.
          in: query
          required: false
          schema:
            type: string
        - name: contact_details
          in: query
          description: >-
            Filter signals according to the presence of contact details, accepts
            (a combination of) the following values `none` for no contact
            details, `phone` for the presence of a reporter's phone number, and
            `email` for the presence of a reporter's email address.
          schema:
            type: string
            enum:
              - none
              - phone
              - email
          required: false
        - name: priority
          in: query
          description: >-
            Filter signals according to their priority, accepts (a combination
            of) the following values: `low`, `high` and/or `normal`.
          required: false
          schema:
            type: string
            enum:
              - low
              - normal
              - high
        - name: feedback
          in: query
          description: >-
            Filter Signals/meldingen according to wether feedback was received
            for them. This filter accepts one of the following values:
            `satisfied` (reporter was satisfied with the resolution),
            `not_satisfied` (reporter was not satisfied with the resolution),
            or `not_received` (no feedback was sent by the reporter). Only the
            newest feedback is taken into account.
          required: false
          schema:
            type: string
            enum:
              - satisfied
              - not_satisfied
              - not_received
        - name: source
          in: query
          description: >-
            Filter Signals/meldingen according to their source, accepted values
            are derived from the database. Source `online` are Signals/meldingen
            that were created in the public website.
          required: false
          schema:
            type: string
        - name: 'type'
          in: query
          description: >-
            Filter signals according to the currently assigned type, accepts
            (a combination of) the values `SIG` a "Signal/Melding", `REQ` a "Request/Aanvraag", `QUE` a "QuestionVraag", `COM` a "Complaint/Klacht" and `MAI` a "Maintenance/Groot onderhoud"
          schema:
            type: string
            enum:
              - SIG
              - REQ
              - QUE
              - COM
              - MAI
          required: false
        - name: note_keyword
          in: query
          description: >-
            This filter option is temporary, and will likely be replaced by
            either proper tagging or extended search using the Elastic Search
            implementation of SIA.
            Filter Signals/meldingen according to the presence of a certain
            keyword in one or more of notes associated with a Signal/melding.
            Accepts one keyword, will check wether it is present and return the
            Signals/meldingen where it is present.
          required: false
          schema:
            type: string

      responses:
        '200':
          description: List of signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1PrivateSignalListGetGeography'
        '401':
          description: Not authenticated, may be caused by expired token.
        '403':
          description: Not authorized to access this endpoint.
      security:
        - OAuth2:
            - SIG/ALL
...
  /signals/v1/private/signals/{id}/children/:
    parameters:
      - name: id
        in: path
        description: ID of signal whose child signsl are requested
        required: true
        schema:
          type: integer
          example: 1
    get:
      description: Retrieve child signals by parent signal ID.
      responses:
        '200':
          description: Child signals associated with the specified parent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1PrivateSignalchildrenGet'
        '401':
          description: Not authenticated, may be caused by expired token.
        '403':
          description: Not authorized to access this endpoint.
        '404':
          description: >-
            Parent signal not found or requesting children of a parent signal
            that has no child signals.
      security:
        - OAuth2:
            - SIG/ALL
  /signals/v1/private/signals/{id}/history:
    parameters:
      - name: id
        in: path
        description: ID of signal whose history needs to be retrieved.
        required: true
        schema:
          type: integer
    get:
      description: History of signal instance with specified ID.
      responses:
        '200':
          description: List of history entries for given signal instance.
        '401':
          description: Not authenticated, may be caused by expired token.
        '404':
          description: Signal not found.
      security:
      - OAuth2:
        - SIG/ALL
...
components:
  schemas:
    StatusStateChoices:
      description: Status string (predefined).
      type: string
      enum:
        - m
        - i
        - b
        - h
        - o
        - a
        - s
        - reopened
        - ingepland
        - ready to send
        - sent
        - send failed
        - done external
...
    V1PrivateSignalListGetGeography:
      description: Geography
      type: object
      properties:
        'type':
          type: string
          example: "FeatureCollection"
        features:
          type: array
          items:
            type: object
            properties:
              'type':
                type: string
                example: "Feature"
              geometry:
                type: object
                properties:
                  'type':
                    type: string
                    example: "Point"
                  coordinates:
                    type: array
                    items:
                      type: integer
                    minItems: 0
                    maxItems: 2
                    example: [4.890986949631043, 52.41875227274259]
              'properties':
                type: object
                properties:
                  id:
                    description: The ID of the Signal
                    type: integer
                    example: 1
                  created_at:
                    type: string
                    format: date-time
                    example: "2020-01-01T00:00:00+00:00"
...
    V1PrivateSignalchildrenGet:
      description: >-
        Child signals associated with a specific parent Signal. If the
        requesting user has access to the parent signal and not the child
        signals, they will still be visible in this endpoint in an
        abridged representation.
      type: object
      properties:
        _links:
          description: HAL JSON links
          type: object
          properties:
            self:
              type: object
              properties:
                href:
                  description: URI of these child signals in API
                  type: string
                  format: uri
        count:
          description: Number of child signals associated with parent signal.
          type: integer
        results:
          description: List of child signals
          type: array
          items:
            type: object
            properties:
              _links:
                description: HAL JSON links
                type: object
                properties:
                  self:
                    type: object
                    properties:
                      href:
                        description: URI of this child signal
                        type: string
                        format: uri
              id:
                description: child signal ID
                type: integer
              status:
                description: child signal status
                type: object
                properties:
                  state:
                    type: string
                    description: status "state" code
                  state_display:
                    type: string
                    description: display name of status
              category:
                description: Main and sub category of child signal
                type: object
                properties:
                  sub:
                    type: string
                    description: sub category name
                  sub_slug:
                    type: string
                    description: sub category slug
                  departments:
                    type: string
                    description: department codes associated with sub category
                  main:
                    type: string
                    description: main category name
                  main_slug:
                    type: string
                    description: main category slug
...
  securitySchemes:
    OAuth2:
      description: SIA API is using OAuth2 implicit grant flow.
      type: oauth2
      flows:
        implicit:
          authorizationUrl: '/oauth2/authorize'
          scopes:
            SIG/ALL: General access to SIA Django application.